requirements.txt
python-telegram-bot==13.15
requests==2.28.2
schedule==1.1.0
python-dotenv==1.0.0 